Changed the graph design for the meeting tomorrow
I think this display offers a better sense of the distribution of values and patterns than other specifications I tested.  I tried a ribbon plot that you saw last week, as well as a box-plot with errors variance, but neither really showed was was happening to create the range as well as this plot.